chore: Update GitHub runner version in Dockerfile
Update the GitHub Actions runner version in the Dockerfile
from 2.323.0 to 2.328.0 to ensure that the latest features
and security updates are included. This change is aimed at
keeping the CI/CD processes aligned with the latest stable
releases and improvements provided by GitHub. The ARG
declaration for RUNNER_VERSION is modified as follows:

```dockerfile
- ARG RUNNER_VERSION=2.323.0
+ ARG RUNNER_VERSION=2.328.0
```